Initial File Structure and makefile Made
I have made started file structure and created a makefile that should work for the homework.

Any commands from pgmProcess.c and pgmUtility.c are just stubbed out for the time being.  main.cu is also created and has a message to make sure it was working.

I also moved any files that were from the zip file; however, didn't seem like it would be directly used for the homework or examples are moved into a resource folder.